requirements-dev.txt
# This requirements.txt file is part of the OCMovies-API project and the
# requirements listed are to be installed when developing the API. 
# 

-i https://pypi.org/simple
-r requirements.txt
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
coverage==5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
factory-boy==3.1.0; python_version >= '3.5'
faker==5.0.1; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
packaging==20.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
pytest-mock==3.3.1
pytest==6.2.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
